refactor for better separation
